feat: add "Keep transformation" to menu
This could be helpful when viewing a gallery with similar size, which
need to apply same transformation to the view. E.g. When viewing a batch
of pixel image sprite assets, the view might need to be zoomed to clearly
view the detail of the image.

Other use-case can be seen at the linked issue below.

https://github.com/BLumia/pineapple-pictures/issues/87
https://github.com/BLumia/pineapple-pictures/issues/89